[lib] split os/stream os/formatter

+++ src/lib/base/strings.cpp
@@ -897,8 +897,7 @@ bool _xf_split_first_(const string &s, string &pre, string &f, string &post) {
 
 
 // convert a string to an integer
-int string::_int() const
-{
+int string::_int() const {
 	bool minus = false;
 	int res = 0;
 	for (int i=0; i<num; i++) {
@@ -913,10 +912,9 @@ int string::_int() const
 	return res;
 }
 
-long long string::i64() const
-{
+int64 string::i64() const {
 	bool minus = false;
-	long long res = 0;
+	int64 res = 0;
 	for (int i=0; i<num; i++) {
 		char c = (*this)[i];
 		if ((c == '-') and (i == 0))
@@ -929,18 +927,17 @@ long long string::i64() const
 	return res;
 }
 
-int s2i(const string &s)
-{	return s._int();	}
+int s2i(const string &s) {
+	return s._int();
+}
 
 // convert a string to a float
-float string::_float() const
-{
+float string::_float() const {
 	return (float)f64();
 }
 
 // convert a string to a float
-double string::f64() const
-{
+double string::f64() const {
 	bool minus = false;
 	int e = -1;
 	double res = 0;
@@ -969,19 +966,21 @@ double string::f64() const
 }
 
 
-float s2f(const string &s)
-{	return s._float();	}
+float s2f(const string &s) {
+	return s._float();
+}
 
-double s2f64(const string &s)
-{	return s.f64();	}
+double s2f64(const string &s) {
+	return s.f64();
+}
 
-bool string::_bool() const
-{
+bool string::_bool() const {
 	return (*this == "true") or (*this == "yes");
 }
 
-bool s2b(const string &s)
-{	return s._bool();	}
+bool s2b(const string &s) {
+	return s._bool();
+}
