+++ src/lib/base/strings.h
@@ -116,7 +116,7 @@ class string : public bytes {
 	Array<int> utf8_to_utf32() const;
 	Array<int> utf16_to_utf32() const;
 	int _cdecl _int() const;
-	long long _cdecl i64() const;
+	int64 _cdecl i64() const;
 	float _cdecl _float() const;
 	double _cdecl f64() const;
 	bool _cdecl _bool() const;
